+++ CHANGELOG.rst
@@ -9,7 +9,8 @@ Develop
 
 Major Features and Improvements
 -------------------------------
-- add Landau distribution
+- add ``JohnsonSU`` PDF, the Johnson SU distribution.
+- add Landau distribution.
 
 Breaking changes
 ------------------